Bug 797416 - Enter key does not move to blank tx or next line
Caused by a previous commit that did not allow for just pressing the
enter key without changes to move to the next transaction. This was
part of a change to stop being asked to create a new account many times
when the transfer cell had been changed and declining it. To fix this
when gnc_split_reg_record returns false, test for the current cursor
changed flag being not set to proceed to do the move.