Bug 797397 - Import Map Editor, Ctrl-A with filter
If you use a filter and then use Ctrl-A to select all rows for deletion
all entries are deleted. This is due to the use of a tree store and the
Ctrl-A selecting the parent rows also.

To fix this a tree view 'select' function is used that does not allow
the parent rows to be selected with a filter.