[aqbanking] Bug 797430: Avoid double init of gui
Patch by Mike Pieper. Thanks!

Explanation for why this is necessary: The previous (old) behaviour
changed in 62f7aa66. Before this commit (or when AQBANKING6 is not
set) it calls AB_JobSepaTransfer_new. For AQBANKING6 it calls
AB_Transaction_new. AB_Transaction_new is not using gui. Therefore
gui->state stays on INIT, while before it made several status changes
(=> RUNNING, FINISHED, HIDDEN). On second call to gnc_GWEN_Gui_get it
checks the state and finds it still in INIT which seems to be wrong.

So the right fix is to either use gui in AB_Transaction_new or to
remove first allocation, which is done here.