Bug 759005 - Print negatives in red
Fixes display of negative monetary-amounts so that they are rendered
according to style-sheet option.

Note this commit fixes via <td> renderer -- if it has a single datum, and
has a negative monetary amount, then its tag gets "-neg" appended.

If a gnc:monetary is renderer *outside* html-table-cell, it will not
be rendered as red (in default stylesheet).