Bug 797453 - Chart of Accounts is slow to update / redraw
The chart of Accounts tree view gets redrawn often and gets the data
from the model via gnc_tree_model_account_get_value(). As such the
numeric values are constantly being recalculated on every draw which is
not ideal. This change caches all the string values to a GHashTable for
retrieval which should be quicker than before.
Changes to an account will clear the cache for that account and its
parents so values can be updated. Changes to the preference negative
colour clears the whole cache.