[dialog-payment] tooltip warning if payment is unattached
Completing a payment for a customer without associating with an open
invoice is allowable and is used for prepayments.

Previously this was done silently.

This commit will enable a warning to inform that the payment does not
have an invoice/bill document attached. This warning does not block
the payment however.