Bug 797521 - Receivable Aging (beta): include earliest split
When processing invoice->payments, we used gnc-lot-get-earliest-split
to find the lot opening-split to skip it.

However it does not always return the lot opening-split, because a
lot-reducing split may be dated earlier than the lot opening-split.

This change means that the posting-split is processed together with
the payment splits, but its PaymentAcctSplitList is null, so it does
not add any row to the result.