Take account of some columns in the register having 'Tot' added.
In the journal register and when using auto split some register columns
have 'Tot added to them like 'Tot Debit' depending on where the cursor
is so if you tried to size that column with a double click on the
header it would be based on the short name. This changes fixes that.