Bug 796530 - [txn csv importer] usability suggestions
Fix account selection by typing in the base account combo box text field
The typing triggered the combobox' changed event. That triggered a repopulation
of the csv data treeview (due to possibly having to unset an account column).
That then in the end retriggered setting the base account. However while typing
in the combo box there may not be a valid account selected in the combo box.
So break this short circuit by testing for a change in the last-known base account
compared to what the combo box believes is the proper account. This last-known
account nore the combo box' internal state will change while typing allowing
proper user input.