create a more consistent aqbanking online_id
The online_id will generated based on bankcode and accountnumber. The
accountnumber may have leading zeros which shall be ignored to make the
creation reliable. In my case the accountnumber of the accountinfo has
no
leading zeros, but the accountnumber of the imported transactions has
leading
zeros. In the result the generated online_id was different and the
widget
to assign the transactions to a account appears always.