[trep-engine] upgrade to include 'balance brought forward'
This fairly complex-looking change is actually simple.

default-calculated-cells is redefined so that the running-balance
column has a custom subheading renderer. The subheading renderer was
formerly used solely for the account->friendly-debit/credit-string
renderer, and is now upgraded to handle hard-coded symbols.

If the friendly-heading-fn is a symbol 'bal-bf, it'll print the
split->account->balance at begindate.

Other columns may also be upgraded for novel subheading renderers. No
forward/backward compatibility issues are expected at all. This change
looks complex because the begindate is not available to
make-split-table and has to be passed on as an argument to be
available for use, and the subheading-renderer will test whether to
display the friendly-fn-renderer at a later stage than previously.