Account matching: Test for full and partial online-id matches.
An OFX file import use-case includes a master account with sub accounts
having additional qualifiers appended. 7853f5a2 broke that because the
trailing characters aren't noise.

Instead return only perfect matches (less a trailing space if present,
replacing a second traversal) and cache partial matches and continue the
search. If we find a longer partial match replace the shorter one. If we
find a second account with the same online-id raise an error. If at the
end there's a unambiguous partial match but no full match then it's
either trailing noise from AQBanking or it's a new account. We use the
passed-in new_account_default_type to decide which: If it's
ACCT_TYPE_NONE then we use the partially-matched account, otherwise we
continue on to handle the remaining arguments.