Initially select the last account chosen in the account picker dialogs.
Separately keeps track of last investment account, security account,
and income account.  One issue is that gnc_import_select_account
doesn't tell the caller if it put up a dialog or found the online ID
on an existing account.  This means the last account may be one the
user didn't manually select.  This may or may not be the right thing
to do.