Change the way the focus is added to plugin_pages
This change corrects a previous commit and makes all the plugin_pages
follow the same format. In the previous commit a test was made for the
plugin_page to be equal to one returned from get_current_page which
would always be the case. With this change the respective plugin_page
is passed as a parameter to the 'page_changed' call back and it is this
that is tested against the current plugin_page.