Bug 797522 - Focus after reconcile jumps to a different account
After the use of the reconcile window, the keyboard focus may not be
on the current account register. This is due to the call for a gui
refresh which refreshes the pages in reverse qofinstance order and as
part of this there is a call to grab the keyboard focus on the register
sheet so the last one will have the keyboard focus which may not be the
current register, just depends on the order of creation.

To fix this, as part of the main window page changed callback a flag is
set on the current sheet to indicate it can grab the focus on refresh.