[lot-viewer.scm] initial commit
Simple report to visualise lots (business and non-business) in a
spreadsheet. Each column is a unique lot, and each row is a unique
transaction. The lot balance, and the associated lot invoice are also
included.

Each transaction is analysed to render the split in the lot column,
and non-lot splits are rendered in a last column.

An account must be chosen in options, and all splits in the specified
date range will be scanned to find unique lots which are all reported
in order of first appearance.

A heavy APAR account will lead to an impractical number of columns
signifying multiple lots, therefore date-range filtering and
description filtering are both encouraged to narrow down the list of
transactions.