Bug 797570 Import Customers/Vendors, Bills/Invoices: dialog with not imported rows does not expand with window size
In the import of Customer/Vendors en Bills/Invoice, the dialog that lists rows that were not imported shows a scrollable view of only four rows.
The view does not use the available space of the window, and does not expand with resizing the window. This change fixes this issue.
In the same functionalities, the dialog for a user defined regular expression has a similar issue. This changes also fixes that issue.