Bug 796911 - Minimum window width to large.
This is down to the amount of information that is displayed on the
register status bar which can also be influenced by the type of
register being displayed. To fix this the text labels used have been
enabled to ellipsize at the end and also the displayed information has
been added to a tooltip. So for example the minimum app size was
957x736 and after the changes it can be 610x475.