Reports were not being loaded
The reports page uses a g_idle_add against the plugin_page to load the
report once the container for the report is realized. With the changes
to the page focus functions, the use of g_idle_remove_by_data removed
this idle function so no report. Change the page focus functions to
record the id used and then use this id to remove the page focus idle
function.