[html-chart] compact, use (ice-9 match)
Use (ice-9 match) for easier matching for the variable
'path'. Explanation; consider the snippet

(cond
 ((null? path) ...A...)
 ((and (pair? (car path)) (number (caar path))) ...B...)
 (else ...C...))

This snippet is a shorthand for nested if-then-else clauses, testing
'path' successfully to determine whether to evaluate A, B or C. Some
code will also use components of 'path' eg B uses (caar path).

Using Alex Shinn's match.scm library allows more concise matching and
assignment at the same time. The syntax is (match EXPR (CLAUSE BODY ...))

(define (try path)
 (match path
   (() (display "null"))
   ((((? number? idx)) . tail) (display "B") (display idx) (newline) (display tail))
   ((head . tail) (display "C") (display head) (newline) (display tail))))

A: the first match is easy -- if path is '() then evaluate the first
body.

     (try '())
 --> "null"

C: the third match is easy -- if path is a pair, then assign 'head' to
pair's car, 'tail' to the pair's cdr, and evaluate the body which has
access to head and tail. Note the head is a string, and the tail is a
list containing a single string.

     (try '("this" "that"))
 --> Cthis
     (that)

B: the second match is more difficult -- let's consider the
broken-down clause: a pair, (HEAD . tail); where HEAD is a
single-element list (ELT), and ELT is a match conditional satisfying
number? and is also assigns the variable idx (? number? idx).

Example:
    (try '((2) "two"))
--> B2
    (two)

This means the match is successful when 'path' is a pair, the pair's
car is a single-element list, and the list's sole element is a
number. The latter is bound to the variable 'idx' which is accessible
in the body. The variable 'tail' contains the path's cdr which
contains a single string.

Note: later in same commit we also use the identifier _ to denote
elements which *must* be matched, but are *not* bound to any variable.

e.g. the clause (((? out-of-bound?)) . _) means that path is a pair,
whose car is a single-element list, and the element satisfies the
predicate out-of-bound?. We don't need to use the (cdr path) therefore
we use _ as a placeholder.