Clear hbci KVP from accounts no longer associated with an online one
A reverse hash table is introduced in order to memorize the
matched GnuCash accounts before starting the matching wizard.
This hash table is used within the aai_on_finish callback
to delete only the KVPs of those GnuCash accounts which are
no longer matched with an AqBanking account.
All other GnuCash accounts (previously matched and currently
matched) are just updated with the new assignments.