L12N:de_DE: SKR04: Part 2 of Bug 513000
Caused by an error in the german translation in 2005 I had created many
accounts of type PAYABLE, which should have been normal LIABILITY
accounts. They should not show up in the "Post [invoice] to account"
list.

Because the other german contributors in 2008 did not understand the
difference between the account types, the change had been rejected for
SKR04, although it was accepted for SKR03.