Correct check amount when printed from subaccount
Previously, GnuCash would include only a single split's amount as the check
amount. If a "bank" account has multiple subaccounts for budgeting reasons, and
a transaction involves both "internal" (e.g. from one bank subaccount to
another bank subaccount) and external transfers, then even when a "subaccount"
register showing the top-level bank account is printed from, the check will
have the amount of the first split rather than the total of all splits which
are from subaccounts.

This change communicates the parent account (when viewed from a "subccount"
register) to the check printing dialog so that the check amount matches the
amount displayed in the register.