Revert "completion redefine gnc:debug" and fix some bugs in it
This change made it impossible to turn debug output on or off on the
fly by calling qof-log-set-level from Scheme code.  The optimization
achieved isn't all that great either since the arguements to
gnc:debug are still evaluated when debugging is off and this is
where a lot of the overhead is. Even without this change the call to
strify is avoided.

Also fixed the parameters of qof-log-check:
    "gmc" => "gnc.scm"
    G-LOG-LEVEL-DEBUG => QOF-LOG-DEBUG

This reverts commit b3a4cd62775705b2543438f912ca71dad591bf0e.