Bug 797631 - Superfluous account selection dialog on first online retrieval
The online id, which is needed to find a GnuCash account for a transaction
or the balance retrieved from an online account, is already assigned to each
matched GnuCash account within the Online Banking Setup tool.
The online id is removed from the GnuCash account if it is no longer matched
with an AqBanking account.