Bug 797638 - SIGSEGV editing transaction after closing Find Accounts dialog.
Crash was from dangling qof-event handler because the dialog was closed
with the title-bar button instead of the close button.

Since the close button destroys the window, simply moving the other
calls in close_handler to destroy_cb causes the titlebar close button
to behave the same as the close button and prevents the crash.

message for your changes. Lines starting