Bug 797481 - crash on close of unsaved tabs by pressing [X]
My original fix for this was a bit off, the problem wasn't that
the sheet had been freed, it was that the item_edit had so the
ptr to the sheet was garbage. The correct solution is to catch
the item_edit's "destroy" signal and remove the idles.

This commit replaces b5fdcfcb5.