Bug 797084 - Provide an option to leave edited transactions reconciled
Instead of making this optional, changes to transaction fields in
case the transaction has reconciled splits will now only issue a
warning. If the user continues anyway with changes to the transaction
field the reconcile state of splits won't be affected.
Changing protected split fields will continue to unreconcile that split.