Add reconcile immediately after ofx import, based on statement
To make this work, I had to add the right head and make a change to the CMakeLists file.

The ofx code used to have a static int to count imported transactions, which was a bit hacky.
I replaced it with a structure that's passed to all the callbacks. The structure has
the transaction count, as well as a list of statement balances information.
This list is used after the import of the ofx file to do the reconciliation.
NOTE: I wouldn't need a list since the current code only process one ofx at a time. However
I have another PR for importing several ofx files in one shot, so this PR is ready for that.
The reconciliation code is modeled after what's done in aqbanking.