Reconcile window - Retain position in split list after deletion
The changes for Bug 797236 so that the last selected split is
visible when the list is refreshed has no effect when the Delete
button is used, as the selected split has been deleted.
So preselect the next split after the split to be deleted.
If no next split in the list, select the previous.
The split preselected must have a different parent transaction to
the split to be deleted, as all splits in the transaction will be
deleted.