Don't mutate the cached global account list store.
That defeats the purpose of caching it and causes problems when using it
from two registers at once. Instead use the PopBox tmp_store for listing
the type-ahead match results if any.

Since a no-match causes an empty ItemList and an empty ItemList with
grab loses key events because there's no cell to handle them--very
confusing to the user--prevent that by checking that there are entries
before grabbing.