Bug 797730 - Transaction matching can match multiple imported transactions to the same existing one
This is a simple solution to the conflict problem. It does not guarantee an optimal match (i.e. a match with
the best combined match score) but it removes conflicts and finds an acceptable solution. The idea is simple:
- Start with the first imported transaction,
- Find all other imported transations that conflict with it (i.e., want to match the same register transaction).
- Of all these conflicting transactions, find the one with the best match to the register transaction
and set it as the winner: it keeps its match, and all other transactions go to their next favorite match.
- If any conflict happened, start back from the top as new conflicts can arise when imported transactions
lose their top match and go to their second best one (if any).

This loop is guaranteed to terminate because every time we go to the top, we remove at least one match
from the sum of all matches.

A bit of cleanup, a few questions are unresolved

Bug 797730 - Transaction matching can match multiple imported transactions to the same existing one

This is a simple solution to the conflict problem. It does not guarantee an optimal match (i.e. a match with
the best combined match score) but it removes conflicts and finds an acceptable solution. The idea is simple:
- Start with the first imported transaction,
- Find all other imported transations that conflict with it (i.e., want to match the same register transaction).
- Of all these conflicting transactions, find the one with the best match to the register transaction
and set it as the winner: it keeps its match, and all other transactions go to their next favorite match.
- If any conflict happened, start back from the top as new conflicts can arise when imported transactions
lose their top match and go to their second best one (if any).

This loop is guaranteed to terminate because every time we go to the top, we remove at least one match
from the sum of all matches.