Bug 797689 - Change the Bill Terms Table dialogue to be based on a  GtkWindow
When using Gnucash on Microsoft Windows and a dialog opens a further
dialog when you switch away from the application and back the last
dialog that has the focus is behind the parent and it appears like the
application has froze. To fix this try to use only one level of dialogue
so change the Bill Terms Table to use a Window with no transient parent.