Enable the Options Checkbox label to be used to toggle
With the new the options dialogue using a GtkGrid it is required that
clicking on the label of the GtkCheckBox would also toggle the
GtkCheckBox. To this end the label is added to an event box so it can
be clicked on and a callback used to toggle the checkbox.