Implement multiple-ofx import with reconcile
I had to rebase against master which already had the reconcile after ofx import, so this took some effort.
At this point, this PR implements the ability to import multiple ofx files, one after the other.
For each, you have to option of triggering a reconcile after the match, available as a check box
at the bottom of the match dialog.