Move the transaction association dialog to dialog-assoc.c/h
The transaction association dialog was all setup from gnc-split-reg but
it seems logical to move all the association source to one file so this
commit does that as well as changing the dialog to be based on a glade
file. The existing toolbar menu entries have changed to Update, Open and
Remove the association.