Display glyphs for Transaction Associations
Display glyphs in the register for transaction associations if a font
has them. A test is made when the register is loaded to see if there is
a font that has the two glyphs available otherwise the usual characters
are used. The two glyphs used are...

GLYPH_PAPERCLIP  // Codepoint U+1F4CE
GLYPH_LINK       // Codepoint U+1F517