Bug 797754 - Account Register scrolling stops working
In an account register, scrolling stops working when the mouse pointer
re-enters the scrollbar after leaving it. This was caused by a fix for
the scrollbar not being updated in earlier versions of Gtk+. Having
removed this earlier fix and tried on Windows 10 and two flavours of
Linux with the current version of Gtk+ the earlier problem is not
evident so this commit removes that fix.